+++ kedro_vertexai/__init__.py
@@ -1,3 +1,3 @@
 """kedro_vertexai."""
 
-version = "0.3.0"
+version = "0.4.0"